Refresh token - Stockfinder
Refresh tokens are now handled correctly. This means that the
application is keeping track of the refresh token in a configuration
file.

Stockfinder helper application is now started and uses the same
configuration file to handle the refresh token.

Started working on lua requires so that both applications aren't located
in one hudge file, and in hopes that functions can be reused for both
applications.